Refactoring in progress (Norme 42) : printf_unicode.c

+++ Includes/Print_Unicode/print_unicode.h
@@ -5,8 +5,8 @@
 #include "create_and_print_t_bin_list.h"
 #include "../Handlers/handlers.h"
 
-size_t			my_put_wint_t(int dec, t_flag flag);
-int			my_put_wchar_t(wchar_t *string_wchar, t_flag flag);
+intmax_t			my_put_wint_t(int dec, t_flag flag);
+intmax_t			my_put_wchar_t(wchar_t *string_wchar, t_flag flag);
 t_bin_list			*call_functions_to_convert_dec_to_bin_in_utf8(int dec, int number_of_bytes[1]);
 
 #endif
